Separate regions generation into separate module
