+++ iCount/genomes/constants.py
@@ -1,0 +1,69 @@
+"""Constants."""
+from collections import OrderedDict
+
+TYPE_HIERARCHY = [
+    'CDS',
+    'UTR3',
+    'UTR5',
+    'ncRNA',
+    'intron',
+    'intergenic',
+]
+
+SUBTYPE_GROUPS = OrderedDict([
+    ('mRNA', [
+        'IG_C_gene',
+        'IG_D_gene',
+        'IG_J_gene',
+        'IG_LV_gene',
+        'IG_V_gene',
+        'TR_C_gene',
+        'TR_D_gene',
+        'TR_J_gene',
+        'TR_V_gene',
+        'non_stop_decay',
+        'nonsense_mediated_decay',
+        'polymorphic_pseudogene',
+        'protein_coding',
+        'retained_intron',
+        'sense_intronic',
+        'sense_overlapping',
+        '3prime_overlapping_ncRNA',
+    ]),
+    ('lncRNA', [
+        'IG_C_pseudogene',
+        'IG_J_pseudogene',
+        'IG_V_pseudogene',
+        'IG_pseudogene',
+        'TR_J_pseudogene',
+        'TR_V_pseudogene',
+        'TEC',
+        'antisense',
+        'antisense_RNA',
+        'bidirectional_promoter_lncRNA',
+        'lincRNA',
+        'ncRNA sRNA',
+        'non_coding',
+        'macro_lncRNA',
+        'misc_RNA',
+        'misc_RNA_pseudogene',
+        'processed_pseudogene',
+        'processed_transcript',
+        'pseudogene',
+        'transcribed_processed_pseudogene',
+        'transcribed_unitary_pseudogene',
+        'transcribed_unprocessed_pseudogene',
+        'translated_processed_pseudogene',
+        'unitary_pseudogene',
+        'unprocessed_pseudogene',
+    ]),
+    ('miRNA', ['miRNA', 'miRNA_pseudogene']),
+    ('mt_tRNA', ['Mt_tRNA', 'Mt_tRNA_pseudogene']),
+    ('mt_rRNA', ['Mt_rRNA']),
+    ('rRNA', ['rRNA', 'rRNA_pseudogene']),
+    ('sRNA', ['sRNA', 'ribozyme', 'scRNA', 'scRNA_pseudogene', 'vaultRNA']),
+    ('snRNA', ['snRNA', 'snRNA_pseudogene']),
+    ('snoRNA', ['snoRNA', 'snoRNA_pseudogene', 'scaRNA']),
+    ('tRNA', ['tRNA_pseudogene']),
+    ('intergenic', ['intergenic']),
+])
